Fix for #1
Creation of a basic data structure.  Note: This is oversimplified just
so that I can test out a POC.  The actual data structure needed to
support the project I have in mind is definitely more complex than
this, but I want to get to the PHP portion of the POC as soon as
possible.